chore: revisão de código e substituição de DTO

+++ ms-vendedor/src/main/java/com/simoes/ms_vendedor/entity/Pedido.java
@@ -16,8 +16,8 @@ public class Pedido {
     private Long idUsuario;
     private String item;
     private int quantidade;
-    private BigDecimal valor;
-    private BigDecimal valorTotal;
+    private double valor;
+    private double valorTotal;
     private String status;
     private boolean aprovado;
     private String observacao;
