+++ ms-vendedor/src/main/java/com/simoes/ms_vendedor/entity/PedidoProcessado.java
@@ -17,8 +17,8 @@ public class PedidoProcessado {
         private Long idUsuario;
         private String item;
         private int quantidade;
-        private BigDecimal valor;
-        private BigDecimal valorTotal;
+        private double valor;
+        private double valorTotal;
         private String status;
         private boolean aprovado;
         private String observacao;
